Use std::isnan() and isinf() with C++11 (Trilinos #239)
The ATTB machine 'white' which is a POWER8 with GCC 4.9.2 and CUDA 7.5 fails
the unit tests for ST::isnaninf() which uses the genetic inequality test for
NaNs.  This may be due to heavy compiler optimizations.  Therefore, we are
going to try to C++11 functions std::isnan() and std::isinf().  I have ifdefed
this based on C++11.  Without C++11, it just uses the old generic
implementaion.

I have tested this with and without C++11 enabled and they both passed on the
machine hansen using the GCC 4.8.4 compiler.