Add yaml-cpp to standard SEMS env and CI build (#1167)
These changes add the SEMS-provided yaml-cpp TPL to the Trilinos SEMS Dev Env
and the standard CI build.

Unfortunately, this fails for the current standard CI build with
load_ci_sems_dev_env.sh because SEMS did not build the yaml-cpp TPL for GCC
4.7.2.

But this works fine for GCC 4.8.4, for example.