Allow setting exact TPL_BoostLibs_LIBRARIES (TRIL-212)
The SPARC TPLs only install *.a files, even for shared library builds.
Therefore, to support that, we need to update that ATDM Trilinos
configuration system to allow set TPL_BoostLibs_LIBRARIES explicilty.

But the default is still to just use BOOST_LIBS and use *.so or *.a
according to BUILD_SHARED_LIBS.