Switch from intel-18.0.2 to intel-18.0.5 (#3499)
This fixes the problem with the MKL GEEV function that was breaking the test
Anasazi_Epetra_BKS_norestart_test_MPI_4.  Note that I had to upgrade the
compiler and MKL from 18.0.2 to 18.0.5.  I tried just updating MKL but I got
link failures.

I also stop loading the module cmake/3.9.0 since I am manually setting PATH to
point to cmake 3.11 that I installed for 'mutrino'.

NOTE: This change does not fix the test Anasazi_MultiVecTraitsTest2_MPI_4 also
reported in #3499.  That is a seprate problem.