Xpetra: Simplify if() statement (#11425)
The old if() logic was not meaningful.  If you want to assert that a varaible
is defined, then call:

  assert_defined(<varName>)

then include it in an if() statement:

  if (<varName>)

What is the value or the purpose of writing an if statement like:

  if (DEFINED <varName> AND ${<varName>})

?

If you misspell the var name it will always be undefined and that if
statemement will always evaluate to false.

This triggered a reconfigure error while working to reproduce #11425.