Factor out member function Tpetra::CrsMatrix::transferAndFillComplete_getCallersParamters()
This is the most straightforward Extract Function refactoring you can do for
this chuck of code.  This chuck of code calls member functions on the class
Tpetra::CrsMatrix object so it could not be a free function.

This is not very complex code, but at the same time, there are a few tricky
details to get it to work.  (I.e. you can't just copy and paste body into a
new function and then add reference arguments for the input and output data.
You have to add using declarations, etc., for the chuck of code to compile and
run.)

This likely took me 1.5 hours to complete once I got set up given the overhead
of working with a large code base.  And this is really not complicated code.

Signed-off-by: Roscoe A. Bartlett <rabartl@sandia.gov>