Factor out transferAndFillComplete_getCallersParamters() as free function
By just passing in the graph, this function can be extracted as a free
function, that could be unit tested independently from the CrsMatrix class.

I moved the declaration out of the decl.hpp file because it does not have to
be part of the class definition anymore.  But if you were going to unit test
this, you would leave the declaration in the decl.hpp file so you could call
it from a unit test harness.

Signed-off-by: Roscoe A. Bartlett <rabartl@sandia.gov>